Run sbs2 *PostProcessor tests in the shared endpoint context
Extends AbstractEndpointTest and drops the per-class @SpringBootTest,
@Import and inner @TestConfiguration in four tests
(ProxyingDataSourceBeanPostProcessorTest,
TransactionMonitoringBeanPostProcessorTest,
ValueInjectionTrackerBeanPostProcessorTest,
QualifiersPersistencePostProcessorTest) so they join the same
ApplicationContext as the *EndpointTest classes instead of bootstrapping
their own. Inner fixtures move out: ValueInjectionTracker beans become
top-level (TestBeanWithCustomAnnotations / TestBeanWithSpEL /
@TimeoutValue) registered via @Bean methods in
SharedEndpointTestConfiguration; QualifiersPersistence fixtures move to
a new QualifiersPersistenceTestFixtures @TestConfiguration that the
shared config @Imports. AbstractEndpointTest's @DynamicPropertySource
now also publishes the two test.* placeholders without defaults that
TestBeanWithCustomAnnotations needs to construct
(test.spring.profiles.active, test.method.timeout).

Profiler confirms: 18 test classes now share one context, up from 14.
