+++ sbs/axelix-spring-boot-2/src/test/java/com/axelixlabs/axelix/sbs/spring/core/beans/QualifiersPersistenceTestFixtures.java
@@ -1,0 +1,130 @@
+/*
+ * Copyright (C) 2025-2026 Axelix Labs
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.axelixlabs.axelix.sbs.spring.core.beans;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+
+/**
+ * Bean declarations that drive {@link QualifiersPersistencePostProcessorTest}. Imported into the
+ * shared endpoint test context so the {@code BeanFactoryPostProcessor} populates
+ * {@link DefaultQualifiersRegistry#INSTANCE} with the test-specific bean names.
+ *
+ * @author Mikhail Polivakha
+ */
+@TestConfiguration
+@Import({
+    QualifiersPersistenceTestFixtures.BeanMethodDeclarations.class,
+    QualifiersPersistenceTestFixtures.ComponentMethodDeclarations.class,
+    QualifiersPersistenceTestFixtures.ConfigurationClassesDeclarations.class
+})
+public class QualifiersPersistenceTestFixtures {
+
+    @Documented
+    @Target({ElementType.TYPE, ElementType.METHOD})
+    @Retention(RetentionPolicy.RUNTIME)
+    @Qualifier("customQualifier")
+    public @interface CustomQualifier {}
+
+    @TestConfiguration
+    static class ComponentMethodDeclarations {
+
+        @Component("noQualifiersBeanName")
+        static class NoQualifiers {}
+
+        @Service("builtInTypeQualifierBeanName")
+        @Qualifier("builtInTypeQualifier")
+        static class BuiltInTypeQualifier {}
+
+        @Service("customTypeQualifierBeanName")
+        @CustomQualifier
+        static class CustomTypeQualifier {}
+
+        @Service("mixedTypeQualifierBeanName")
+        @CustomQualifier
+        @Qualifier("builtInTypeQualifier")
+        static class MixedTypeQualifier {}
+    }
+
+    @TestConfiguration
+    static class BeanMethodDeclarations {
+
+        static class BeanMethodNoQualifiers {}
+
+        static class BeanMethodBuiltInTypeQualifier {}
+
+        static class BeanMethodCustomTypeQualifier {}
+
+        static class BeanMethodMixedTypeQualifier {}
+
+        @Bean("beanMethodNoQualifiersBeanName")
+        public BeanMethodNoQualifiers beanMethodNoQualifiers() {
+            return new BeanMethodNoQualifiers();
+        }
+
+        @Bean("beanMethodBuiltInTypeQualifierBeanName")
+        @Qualifier("builtInTypeQualifier")
+        public BeanMethodBuiltInTypeQualifier builtInTypeQualifier() {
+            return new BeanMethodBuiltInTypeQualifier();
+        }
+
+        @Bean("beanMethodCustomTypeQualifierBeanName")
+        @CustomQualifier
+        public BeanMethodCustomTypeQualifier customTypeQualifier() {
+            return new BeanMethodCustomTypeQualifier();
+        }
+
+        @Bean("beanMethodMixedTypeQualifierBeanName")
+        @CustomQualifier
+        @Qualifier("builtInTypeQualifier")
+        public BeanMethodMixedTypeQualifier mixedTypeQualifier() {
+            return new BeanMethodMixedTypeQualifier();
+        }
+    }
+
+    @Configuration
+    static class ConfigurationClassesDeclarations {
+
+        @Configuration("noQualifiersConfigBeanName")
+        static class NoQualifiers {}
+
+        @Configuration("builtInTypeQualifierConfigBeanName")
+        @Qualifier("builtInTypeQualifier")
+        static class BuiltInTypeQualifier {}
+
+        @Configuration("customTypeQualifierConfigBeanName")
+        @CustomQualifier
+        static class CustomTypeQualifier {}
+
+        @Configuration("mixedTypeQualifierConfigBeanName")
+        @CustomQualifier
+        @Qualifier("builtInTypeQualifier")
+        static class MixedTypeQualifier {}
+    }
+}
